requirements.txt
python-socketio[client]
pypozyx
pyserial
pyyaml